backend/requirements.txt
fastapi==0.135.1
huggingface-hub==1.7.2
llama-cpp-python==0.3.16
pydantic-settings==2.13.1
qdrant-client[fastembed]==1.17.1
uvicorn==0.42.0
